refactor: move component to own folder
