+++ src/components/Header.js
@@ -1,5 +1,3 @@
-const Header = () => {
+export const Header = () => {
   return <p className="header">Mercenary Instruments</p>
 }
-
-export default Header
