+++ src/components/index.js
@@ -1,0 +1,4 @@
+export * from './Header'
+export * from './DisplayContainer'
+export * from './Footer'
+export * from './KeyPad'
